scheduled: Add `createScheduled` primitive
