+++ .changeset/bright-timers-trade.md
@@ -1,0 +1,5 @@
+---
+"@solid-primitives/scheduled": minor
+---
+
+Add `createScheduled` primitive.
